refactor(mesh): apply /simplify review cleanups
- Extract shared <DotPill> in components/detail/status-pill.tsx —
  Task/Escalation/Negotiation pills collapse to 1-line wrappers around
  the same JSX. SessionStatusPill keeps its own shape (different size,
  optional pulse).
- Hoist <MeshBackLink> into components/detail/ — escalation and
  negotiation detail pages now import the single shared component
  instead of declaring identical local consts.
- Move NegotiationNotFoundError from escalation-service.ts to
  negotiation-service.ts (correct dependency direction; escalation now
  imports it from negotiation). routes/escalation.ts and the
  escalation test follow the new path.
- Inline the single-error handleNegotiationError into the route's
  catch block — the indirection wasn't earning its keep with just one
  404 case to map.

Net -41 lines; 21 core unit + 26 api integration tests still green;
web typecheck + lint clean.
