+++ packages/web/app/(authed)/escalations/[id]/escalation-detail-client.tsx
@@ -2,7 +2,8 @@
 
 import Link from "next/link";
 import { useState, type ReactNode } from "react";
-import { ArrowLeft, AlertTriangle, Info, Scale } from "lucide-react";
+import { AlertTriangle, Info, Scale } from "lucide-react";
+import { MeshBackLink } from "@/components/detail/mesh-back-link";
 import { useEscalation } from "@/lib/hooks/use-escalations";
 import { useResolveEscalation } from "@/lib/hooks/use-escalation-mutations";
 import { isApiConfigured } from "@/lib/api/config";
@@ -33,16 +34,6 @@ interface SideView {
   submittedAt?: string;
   recovered?: EscalationRecoveredPosition;
 }
-
-const MeshBackLink = () => (
-  <Link
-    href="/mesh"
-    className="inline-flex items-center gap-1 text-xs text-muted-foreground hover:text-foreground transition-colors mb-3"
-  >
-    <ArrowLeft className="h-3 w-3" />
-    Mesh
-  </Link>
-);
 
 export function EscalationDetailClient({ escalationId }: { escalationId: string }) {
   const { data, isLoading, isError } = useEscalation(escalationId);
